need refactor

+++ get_next_line_bonus.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_next_line_bonus.h                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jleslee <jleslee@student.42.fr>            +#+  +:+       +#+        */
+/*   By: layano <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/11/24 23:16:45 by jleslee           #+#    #+#             */
-/*   Updated: 2021/11/26 23:19:55 by jleslee          ###   ########.fr       */
+/*   Created: 2021/12/20 17:35:39 by layano            #+#    #+#             */
+/*   Updated: 2021/12/20 17:35:41 by layano           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 
 char	*line_num(const char *s);
 char	*ft_strjoin(char *s1, char *s2);
-size_t	ft_strlen(const char *p);
+int		ft_strlen(const char *p);
 char	*get_next_line(int fd);
 
 #endif
